Limit the device state changes which can trigger a recall.
Testing showed that a device which changed from "ringing" to
"not in use" would result in the caller being called back. The
problem is that this does not indicate actual activity by a
human on the monitored side of the call. Now, only state changes
to "not in use" or "unknown" from "unavailable," "busy," or "in use"
will trigger a recall.