CallerID channel DAHDI port FXS are empty after the first call.
The bug is exposed if MFC/R2 support is built into asterisk (i.e.,
openr2.h is present in the include path).  Code that unconditionally
clears the CallerID name and number is included.

Also fixed a malformed if test in mkintf() added by issue 15883.
Converted the if statement to a switch statement for clarity.

Regression of the issue 15883 fix.

(closes issue #16968)
Reported by: grecco
Patches:
      issue16968.patch uploaded by rmudgett (license 664)

(closes issue #16747)
Reported by: viniciusfontes