Don't recreate peer, when responding to a repeated deregistration attempt.
When a reply to a deregistration is lost in transmit, the client retries the
deregistration.  Previously, this would cause a realtime/autocreate peer to be
loaded back into memory, after it had already been correctly purged.  Instead,
we just want to resend the reply without loading the peer.

(closes issue #16908)
 Reported by: kkm
 Patches: 
       20100412__issue16908.diff.txt uploaded by tilghman (license 14)
 Tested by: kkm