Remove usage of `id` since it isn't useful and was causing breakge.
Solaris `id` doesn't support the -u argument.  Instead of figuring out how to
fix this to work on Solaris, I decided to check why it was necessary and where
else it was used.  It was only used in one place, and it hasn't been needed
for a very long time (I question whether it was ever needed).