Internal timing is now on by default, if you're using DAHDI 2.3 or above.
The reason for ensuring DAHDI 2.3 or above is that this version ensures that
a timer is always available, whereas in previous versions, it was possible
for DAHDI to be loaded, but have no drivers to actually generate timing.  If
internal_timing was turned on in this circumstance, a complete lack of audio
would result.  This is the reason why internal_timing was not on by default.
However, now that DAHDI ensures the availability of a timer, there is no
reason for this setting to be off (and in fact, it solves a great many initial
user problems).

(closes issue #15932)
 Reported by: dimas
 Patches: 
       20100519__issue15932.diff.txt uploaded by tilghman (license 14)
 Tested by: tilghman