Eliminate the ast_channel.dialed.number.valid flag.
It is not really needed.  An incoming call should always have a DNID
number.  An outgoing call channel structure does not have a DNID value
unless it is copied from the bridged incoming channel.  The fact that the
dialed number string is a NULL pointer can be used to infer if it is
valid.