Check if DADHI exists before restarting.
New logic added to check if user / group exists, if not we now handle it. 

We now default to 750 (dir) and 640 (file) permissions.  TODO: Explore away to make
this dynamic.

Do not create ASTSBINDIR directory, otherwise we will nuke a running system.

Uncomment [files] section in asterisk.conf.  Since we are locking down asterisk, we might
as well use what we have.