only do magic pickup when notifycid is enabled
A new way of doing BLF pickup was introduced into 1.6.2.  This feature
adds a call-id value into the XML of a SIP_NOTIFY message sent to alert
a subscriber that a device is ringing.  This option should only be enabled
when the new 'notifycid' option is set... but this was not the case.  Instead
the call-id value was included for every RINGING Notify message, which
caused a regression for people who used other methods for call pickup.

(closes issue #17633)
Reported by: urosh
Patches:
      chan_sip.txt uploaded by urosh (license )
      blf_cid_issue.diff uploaded by dvossel (license 671)
Tested by: dvossel, urosh, okrief, alecdavis