tos_sip option was not being set correctly
When tos_sip is used, the tos of the sip socket is only set
correctly if the socket binding changes on a reload.  If the binding
stays the same but the TOS changes, the new tos value would not take
into effect.  This patch fixes that.


(closes issue #17712)
Reported by: nickb