Adding a global setting as well.
First I did not want to do this, as we have a global setting in indications.conf
but then I thought that one would want a separate default setting in chan_sip
from the default Asteirsk one, and have that as a default for all devices. So I gave up
my resistance. Tough story. Simple patch.